chore(dataobj): promote metadata types to hold uint64 instead of uint32 (#15850)

+++ pkg/dataobj/internal/encoding/encoder.go
@@ -158,8 +158,8 @@ func (enc *Encoder) append(data, metadata []byte) error {
 		return nil
 	}
 
-	enc.curSection.MetadataOffset = uint32(enc.startOffset + enc.data.Len() + len(data))
-	enc.curSection.MetadataSize = uint32(len(metadata))
+	enc.curSection.MetadataOffset = uint64(enc.startOffset + enc.data.Len() + len(data))
+	enc.curSection.MetadataSize = uint64(len(metadata))
 
 	// bytes.Buffer.Write never fails.
 	enc.data.Grow(len(data) + len(metadata))
